*** empty log message ***

+++ gearbox/Selector.java
@@ -2,12 +2,13 @@ class Selector {
 	/*@
 	specification Selector {
 		double fst, snd;
-		alias pair = (fst, snd);
+		alias (double) pair = (fst, snd);
 	}
 	@*/
 
 
 }
+
